✨ [feat] Category 컴포넌트 구현
